build: update GitHub Actions
Update actions to their latest versions and uses hashes to pin versions
to prevent tampering. Tags are added as comments and both tags and
hashes will be updated in Dependabot pull requests.

Add `--frozen-lockfile` to `pnpm install`s to make it explicit that this
is what happens during ci.

Use GitHub's own actions to publish to GitHub Pages.