feat: add new workflow
Add a workflow that runs on the first of each month.

The workflow will first check if there is a previous open issue and if
so close it.

Then it will create a new issue, label it, and assign it to the
maintainer group that is responsible for maintenance in this repository.