docs: document .npmrc
Because we want to prevent accidental use of *npm*, we have added
`ignore-scripts=true` to .npmrc. The presence of .npmrc means that pnpm
will also write configuration options to this file instead of to
pnpm-workspace.yaml.

Because of this, configuration options were moved back to .npmrc from
pnpm-workspace.yaml.

As a bonus, this allows us to document the settings.